requirements_fastapi.txt
# FastAPI RAG System Requirements
# Sistema completo migrado do Laravel para FastAPI

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Templates
jinja2==3.1.2
aiofiles==23.2.1

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Processing
pandas==2.1.3
numpy==1.24.4

# AI/ML Libraries
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2

# Google AI
google-generativeai==0.3.2
google-cloud-speech==2.21.0
google-cloud-vision==3.4.4

# OpenAI
openai==1.3.7

# Video Processing
yt-dlp==2023.11.16
ffmpeg-python==0.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
lxml==4.9.3

# OCR
pytesseract==0.3.10
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
